requirements.txt
opencv-python==4.9.0.80
python-dotenv==1.0.1
numpy<2
mediapipe==0.10.21
